+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Added `panic-hook` feature, enabled by default, to conditionally include `console_error_panic_hook`
 - [BREAKING] Hid markdown funnctionality behind optional `markdown` feature
 - Implemented `AsAtValue` for `Option<T>`
+- [BREAKING] Added argument `Option<&Namespace>` to functions `Node::from_html` and `El::from_html`.
+- Added macro `raw_svg!` (#589).
+- Added `browser::dom::Namespace` to `prelude`. 
 
 ## v0.8.0
 - [BREAKING] Rename `linear_gradient!` to `linearGradient!` for consistency with the other svg macros (same with `radial_gradient!` and `mesh_gradient!`) (#377).